Create API-M-Query-PBI
The query in PBI is split into several parts. We need to generate a bearer token with the service credentials. That token is stored and later used for every endpoint to allow us to pull the data. We define the symbols prior to that to optimize the space used and since the raw data received is in JSON, we transform it to tables.